+++ example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -5,10 +5,12 @@
 import FlutterMacOS
 import Foundation
 
+import file_picker
 import screen_brightness_macos
 import xue_hua_video_player
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
+  FilePickerPlugin.register(with: registry.registrar(forPlugin: "FilePickerPlugin"))
   ScreenBrightnessMacosPlugin.register(with: registry.registrar(forPlugin: "ScreenBrightnessMacosPlugin"))
   XueHuaVideoPlayerPlugin.register(with: registry.registrar(forPlugin: "XueHuaVideoPlayerPlugin"))
 }